requirements.txt
python-telegram-bot==21.3
aiohttp==3.9.5
python-dotenv==1.0.1
python-whois==0.9.4
dnspython==2.6.1
aiosqlite==0.20.0
